backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
flask-marshmallow
marshmallow-SQLAlchemy
yahoo_finance_api2
